requirements.txt
absl-py==0.8.1
altgraph==0.16.1
astor==0.8.0
auto-py-to-exe==2.6.6
beautifulsoup4==4.8.1
better-profanity==0.5.0
blis==0.4.1
bottle==0.12.17
bottle-websocket==0.2.9
cachetools==3.1.1
catalogue==0.0.8
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cymem==2.0.3
Eel==0.11.0
future==0.18.2
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googletrans==2.4.0
greenlet==0.4.15
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
imutils==0.5.3
inflect==3.0.2
jaraco.itertools==4.4.2
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-models==0.0.7
Keras-Preprocessing==1.1.0
keras-squeezenet==0.4
keras-utils==1.0.13
Markdown==3.1.1
more-itertools==8.0.0
MouseInfo==0.1.2
murmurhash==1.0.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pafy==0.5.5
pathlib==1.0.1
pefile==2019.4.18
Pillow==6.2.1
plac==1.1.3
playsound==1.2.2
preshed==3.0.2
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
PyAutoGUI==0.9.48
pycparser==2.19
pygame==1.9.6
PyGetWindow==0.0.8
PyInstaller==3.5
PyMsgBox==1.0.7
pyperclip==1.7.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.25
PySide2==5.13.2
PySimpleGUI==4.7.1
PySimpleGUIQt==0.30.0
python-vlc==3.0.7110
pyttsx3==2.71
PyTweening==1.0.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.3
selenium==3.141.0
shiboken2==5.13.2
six==1.13.0
sklearn==0.0
sounddevice==0.3.14
soupsieve==1.9.5
spacy==2.2.3
SpeechRecognition==3.8.1
squeezenet==1.1
srsly==0.2.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
thinc==7.3.1
tqdm==4.40.0
urllib3==1.25.7
wasabi==0.4.2
Werkzeug==0.16.0
whichcraft==0.6.1
wikipedia==1.4.0
WMI==1.4.9
wolframalpha==3.0.1
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0
